Moved `convertToUsableFormat` into Header.tsx
